Cancel the OkHttp Call together with the SwingWorker
It seems like Java's standard Thread interrupts aren't handled as seriously as
Call's own cancel(). For example, Call's execute() will ignore thread interrupts
while it's waiting for a slow server's response. In that case the full read
timeout has to elapse for the execution to end. See
<https://github.com/square/okhttp/issues/6611>.

We introduce our own SendWorker with a cancel() method that cancels both the
underlying SwingWorker and Call.